Skip interpolation along last axis
The final axis of the N-dimensional array is actually just an index over the
image number, with each image corresponding to a different manipulation.
Since this isn't a continuous variable, it makes sense to avoid
interpolating over it. That said, in tests where I interpolate on-grid,
it doesn't really seem to make a practical difference.
So this is more cosmetic than anything else.